fix static quantize
