+++ neural_compressor/torch/quantization/config.py
@@ -396,9 +396,9 @@ class StaticQuantConfig(BaseConfig):
 
     @staticmethod
     def get_model_info(model: torch.nn.Module, example_inputs) -> List[Tuple[str, Callable]]:
-        from neural_compressor.torch.algorithms.static_quant import _get_quantizable_ops_recursively
+        from neural_compressor.torch.algorithms.static_quant import get_quantizable_ops_recursively
 
-        model_info, _, _, _ = _get_quantizable_ops_recursively(model, example_inputs=example_inputs)
+        model_info, _, _, _ = get_quantizable_ops_recursively(model, example_inputs=example_inputs)
         return model_info
 
     @classmethod
